Don't delete admins

+++ db/schema.sql
@@ -9711,8 +9711,18 @@ BEGIN
     LIMIT 1;
 
     IF uID IS NOT NULL THEN
+      IF NOT EXISTS (
+        SELECT user_id
+        FROM Admins
+            WHERE user_id=uID
+            UNION
+            SELECT user_id
+            FROM OrganisationMembers
+            WHERE user_id=uID
+      ) THEN
         DELETE FROM GoogleUserDetails WHERE email=mail;
         DELETE FROM Users WHERE id=uID;
+      END IF;
     END IF;
 END//
 DELIMITER ;
